requirements.txt
# Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
sqlalchemy>=2.0.25

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.2
python-dateutil>=2.8.2

# Monday CRM Integration
requests>=2.31.0

# Environment
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
httpx>=0.26.0
